Add files via upload
This is my subway solution using a basic search that takes into account the subway problem guidelines. There is a known problem with how neighboring intersections are identified based on the current state. Another known error is that a minimum step solution is not calculated as this solution would have to use a list that recognized neighboring intersections based on the current state. Additionally, this does not use coordinates yet nor is it in breadth first search, but there is room to make these changes.